v3: custom lists

+++ app.js
@@ -5,6 +5,8 @@ const date = require(__dirname + "/date.js");
 
 const app = express();
 
+let day = date();
+
 app.set("view engine", "ejs");
 
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -13,31 +15,75 @@ app.use(express.static("public"));
 // change localhost to 127.0.0.1 or won't work
 mongoose.connect("mongodb://127.0.0.1:27017/todolistDB");
 
-// schema
+// schema for today list
 const itemSchema = {
   name: String,
 };
 
-// model
+// model for today list
 const Item = mongoose.model("Item", itemSchema);
 
-// retreive data and render
+//  schema for custom list
+const listSchema = {
+  name: String,
+  items: [itemSchema],
+};
+
+// model for custom list
+const List = mongoose.model("List", listSchema);
+
+// retreive data and render (today list)
 app.get("/", function (req, res) {
-  let day = date();
   Item.find({}).then((data) => {
-    res.render("list", { day: day, tasks: data });
+    res.render("list", { listName: day, tasks: data });
   });
 });
 
 // store new item
 app.post("/", function (req, res) {
-  Item.insertMany([new Item({ name: req.body.task })]);
-  res.redirect("/");
+  let listTitle = req.body.add;
+  if (listTitle == day) {
+    Item.insertMany([new Item({ name: req.body.task })]);
+    res.redirect("/");
+  } else {
+    List.findOne({ name: listTitle }).then((data) => {
+      data.items.push(new Item({ name: req.body.task }));
+      data.save();
+      res.redirect("/" + listTitle);
+    });
+  }
 });
 
 // delete checked item
 app.post("/delete", function (req, res) {
-  Item.findByIdAndDelete(req.body.checkbox).then(res.redirect("/"));
+  let listTitle = req.body.listName;
+  if (listTitle === day) {
+    Item.findByIdAndDelete(req.body.checkbox).then(res.redirect("/"));
+  } else {
+    List.findOneAndUpdate(
+      { name: listTitle },
+      { $pull: { items: { _id: req.body.checkbox } } }
+    ).then((data) => {
+      res.redirect("/" + listTitle);
+    });
+  }
+});
+
+app.get("/:listTitle", function (req, res) {
+  let listTitle = req.params.listTitle;
+
+  List.findOne({ name: listTitle }).then((data) => {
+    if (data) {
+      res.render("list", { listName: listTitle, tasks: data.items });
+    } else {
+      const list = new List({
+        name: listTitle,
+        items: [],
+      });
+      list.save();
+      res.render("list", { listName: listTitle, tasks: [] });
+    }
+  });
 });
 
 app.listen(3000, function () {
